add step 3-5

+++ repository/contactsModel.js
@@ -1,30 +1,62 @@
 const Contact = require('../model/contactShema')
 
 
-const listContacts = async () => {
+const listContacts = async (userId, { page = '1', limit = '5', sub } ) => {
+   const options = { owner: userId };
+  if (sub) {
+    options.subscription = { $all: [sub] };
+  }
+  
   try {
-    
-    const results = await Contact.find({})
-    return results
+    const results = await Contact.paginate(options, {
+      page,
+      limit,
+      populate: {
+        path: 'owner',
+        select: 'email subscription',
+      },
+    });
+    const {
+      docs: contacts,
+      totalDocs: total,
+      page: pageNumber,
+      totalPages: allPages,
+    } = results;
+    return {
+      total: total.toString(),
+      limit,
+      totalPages: allPages.toString(),
+      page: pageNumber.toString(),
+      contacts,
+    };
   } catch (error) {
-    console.log(error)
+    console.log(error);
   }
 }
 
-const getContactById = async (contactId) => {
+const getContactById = async (contactId, userId) => {
   try {
    
-    const result = await Contact.findOne({ _id: contactId })
+    const result = await Contact.findOne({ _id: contactId, owner: userId })
+    .populate({
+      path: 'owner',
+      select: 'email subscription',
+    });
     return result
   } catch (error) {
     console.log(error)
   }
 }
 
-const removeContact = async (contactId) => {
+const removeContact = async (contactId, userId) => {
   try {
     
-    const result = await Contact.findOneAndRemove({ _id: contactId })
+    const result = await Contact.findOneAndRemove({ _id: contactId , owner: userId})
+    
+    .populate({
+      path: 'owner',
+      select: 'email subscription',
+    });
     return result
   } catch (error) {
     console.log(error)
@@ -40,14 +72,18 @@ const addContact = async (body) => {
   }
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   try {
    
     const result  = await Contact.findOneAndUpdate(
-      { _id: contactId },
+      { _id: contactId , owner: userId},
       { ...body },
       { new: true }
     )
+    .populate({
+      path: 'owner',
+      select: 'email subscription',
+    });
     return result
   } catch (error) {
     console.log(error)
